added preprocessor defines. Decided NOT to add include and link dirs
AFAICT CMake does not to have cached variables to specify project-level
include and link dirs. So to set these we need to explicitly call
include_directories() and link_directories() in CMake code. The problem
here is that this call is only retained when it is made after a call to
project(), which CMany is unable to do because its preload file is
executed before project() is called. So no luck here.